requirements.txt
asgiref==3.5.2
attrs==22.1.0
beautifulsoup4==4.11.1
CacheControl==0.12.11
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==2.0.1
crashtest==0.4.1
cryptography==37.0.4
cssutils==2.6.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==4.1.4
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-import-export==3.0.2
django-js-asset==2.0.0
django-model-search==3.0.0
django-modeltranslation==0.18.7
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
dulwich==0.20.46
et-xmlfile==1.1.0
filelock==3.8.0
future==0.18.3
html5lib==1.1
idna==3.4
install==1.3.5
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.16.0
keyring==23.9.3
lockfile==0.12.2
MarkupPy==1.14
more-itertools==8.14.0
msgpack==1.0.4
odfpy==1.4.1
openpyxl==3.0.10
packaging==22.0
pexpect==4.8.0
Pillow==9.3.0
pipenv==2022.12.19
pkginfo==1.9.2
platformdirs==2.5.2
poetry==1.3.2
poetry-core==1.4.0
poetry-plugin-export==1.2.0
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pycparser==2.21
pynliner==0.8.0
pyrsistent==0.18.1
PyYAML==6.0
rapidfuzz==2.13.7
requests==2.28.1
requests-toolbelt==0.9.1
SecretStorage==3.3.3
shellingham==1.5.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.3.0
tomli==2.0.1
tomlkit==0.11.5
trove-classifiers==2022.12.22
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.17.1
virtualenv-clone==0.5.7
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
